The Rust Book 2023 Training
Committing stable baseline relative to work completed in Chapter 2.  Implemented support in the guessing_game for graceful handling of errors with proper processing of the Result object and the related Ok and Err enum variants respectively within a match statement.  Code complete on the guessing_game at this time.  First Rust game complete! :-)