The Rust Book 2023 Training
Committing stable baseline relative to work completed in Chapter 3.  Review of the using the let keyword to effectively re-init variables with the exact same name, but as different types or the same type as needed.  The key note here is that the let prefix allows this to happen.  If you don’t use let but want to reassign a variable then the variable by definition needs to be mut aka mutable annotated at instantiation time.  Committing baseline with both examples visible, with the first example using mut commented out for quick reference.