The Rust Book 2023 Training
Committing stable baseline relative to work completed in Chapter 3.  Review of memory safety aspects with an array and attempting to access a value outside of the bounds of the array.  Rust panics and exits the program with a runtime exception, similar to Java.  Apparently in C or C++ this type of check is manual, and can lead to major security and or run time bugs.  Committing baseline with all examples visible for quick reference.